+++ src/modules/config-editor/config-editor.controller.ts
@@ -1,8 +1,11 @@
 import { Controller, UseGuards, Get, Post, Body, Param, Delete } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { ApiTags, ApiBearerAuth } from '@nestjs/swagger';
 import { ConfigEditorService } from './config-editor.service';
 import { AdminGuard } from '../../core/auth/guards/admin.guard';
 
+@ApiTags('Homebridge Config Editor')
+@ApiBearerAuth()
 @UseGuards(AuthGuard())
 @Controller('config-editor')
 export class ConfigEditorController {