+++ src/modules/platform-tools/docker/docker.controller.ts
@@ -1,8 +1,11 @@
 import { Controller, UseGuards, Get, Put, Body } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { ApiTags, ApiBearerAuth, ApiExcludeEndpoint } from '@nestjs/swagger';
 import { AdminGuard } from '../../../core/auth/guards/admin.guard';
 import { DockerService } from './docker.service';
 
+@ApiTags('Platform - Docker')
+@ApiBearerAuth()
 @UseGuards(AuthGuard())
 @Controller('platform-tools/docker')
 export class DockerController {
@@ -28,12 +31,14 @@ export class DockerController {
     return this.dockerService.restartDockerContainer();
   }
 
+  @ApiExcludeEndpoint()
   @UseGuards(AdminGuard)
   @Get('env')
   getDockerEnv() {
     return this.dockerService.getDockerEnv();
   }
 
+  @ApiExcludeEndpoint()
   @UseGuards(AdminGuard)
   @Put('env')
   updateDockerEnv(@Body() body) {