+++ src/modules/platform-tools/linux/linux.controller.ts
@@ -1,8 +1,11 @@
 import { Controller, UseGuards, Put } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { ApiTags, ApiBearerAuth } from '@nestjs/swagger';
 import { LinuxService } from './linux.service';
 import { AdminGuard } from '../../../core/auth/guards/admin.guard';
 
+@ApiTags('Platform - Linux')
+@ApiBearerAuth()
 @UseGuards(AuthGuard())
 @Controller('platform-tools/linux')
 export class LinuxController {