+++ src/modules/plugins/custom-plugins/homebridge-hue/homebridge-hue.controller.ts
@@ -1,8 +1,11 @@
-import { Controller, UseGuards, Body, Get, Header } from '@nestjs/common';
+import { Controller, UseGuards, Get, Header } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { ApiTags, ApiBearerAuth } from '@nestjs/swagger';
 import { AdminGuard } from '../../../../core/auth/guards/admin.guard';
 import { HomebridgeHueService } from './homebridge-hue.service';
 
+@ApiTags('Plugins')
+@ApiBearerAuth()
 @UseGuards(AuthGuard())
 @Controller('plugins/custom-plugins/homebridge-hue')
 export class HomebridgeHueController {
@@ -13,6 +16,7 @@ export class HomebridgeHueController {
 
   @UseGuards(AdminGuard)
   @Get('/dump-file')
+  @Header('Content-disposition', 'attachment; filename=homebridge-hue.json.gz')
   @Header('Content-Type', 'application/json+gzip')
   async exchangeCredentials() {
     return this.homebridgeHueService.streamDumpFile();