+++ src/modules/status/status.module.ts
@@ -1,12 +1,15 @@
 import { Module } from '@nestjs/common';
+import { PassportModule } from '@nestjs/passport';
 import { StatusService } from './status.service';
 import { StatusGateway } from './status.gateway';
 import { PluginsModule } from '../plugins/plugins.module';
 import { ConfigModule } from '../../core/config/config.module';
 import { LoggerModule } from '../../core/logger/logger.module';
+import { StatusController } from './status.controller';
 
 @Module({
   imports: [
+    PassportModule.register({ defaultStrategy: 'jwt' }),
     LoggerModule,
     PluginsModule,
     ConfigModule,
@@ -15,5 +18,8 @@ import { LoggerModule } from '../../core/logger/logger.module';
     StatusService,
     StatusGateway,
   ],
+  controllers: [
+    StatusController,
+  ]
 })
 export class StatusModule { }